Move Digitization from Simulation to Reconstruction (will be done in reconstruction step)
