+++ Reconstruction/RecTracker/src/components/GeometricTrackerDigitizer.cpp
@@ -17,10 +17,10 @@
 #include "DD4hep/Segmentations.h"
 #include "DD4hep/Volumes.h"
 #include "DDSegmentation/BitField64.h"
-#include "Digitization/FCCDigitizationCell.h"
 #include "GaudiKernel/EventContext.h"
 #include "GaudiKernel/ThreadLocalContext.h"
 #include "RecTracker/ACTSLogger.h"
+#include "RecTracker/FCCDigitizationCell.h"
 #include "datamodel/DigiTrackHitAssociationCollection.h"
 #include "datamodel/PositionedTrackHitCollection.h"
 #include "datamodel/TrackCluster.h"
